Make batch menu scrollable.
